fix code blocks without language tag
for code blocks with no language tag, the `language` variable would be 
set to empty string, which then was passed as `-l` option to `pygmentize`
resulting in an empty code block. The `leader` would also end up as `:`.

To fix it, we are setting `language` to `nil` if the tag is not present.